refactor!: rig-rmcp depends on rig-core only — ManagedToolSink moves to rig-core, MCP tools are portable tools
Option A of the dependency-graph cleanup: the rig-rmcp → rig-agent edge
is gone. rig-core ← rig-rmcp and rig-core ← rig-agent; the facade glues.

- rig-core: tool::managed::{ManagedToolSink, ManagedToolToken} — the
  registry contract for externally managed, refreshable tool sources
  (portable tools in, generation tokens out; last-registration-wins,
  retire-on-disconnect) — and PortableDynamicTool::{with_liveness,
  is_live}.
- rig-agent: ToolServerHandle implements ManagedToolSink;
  DynamicTool::from_portable forwards liveness into ErasedTool::is_live;
  ManagedToolToken is re-exported from rig-core.
- rig-rmcp: no agent feature, no rig-agent dependency (dev-only, as the
  reference sink in tests). McpClientHandler<S: ManagedToolSink>
  registers PortableDynamicTools (with a liveness probe bound to the MCP
  transport). The contextual ErasedTool impl, the rmcp_tool* extension
  traits, the prelude, _meta passthrough from ToolContext and
  raw-result preservation are removed (documented breaks); four tests
  asserting them are removed, and the argument-validation tests drive
  McpTool::execute_mcp directly.
- Facade: rmcp = ["dep:rig-rmcp"]; rig::tool::rmcp unchanged.

Verification: workspace check/clippy clean; rig-rmcp 26+1, rig-agent
568, rig-core 1619 (lib), facade 3394 tests pass; doctests and rustdoc
-D warnings clean; wasm checks pass (rig-core, rig-agent, rig); rig-rmcp
on wasm fails with exactly one error; cargo tree -p rig-rmcp has no
rig-agent, -p rig-agent has no tokio/rmcp.
