+++ Cargo.toml
@@ -365,7 +365,7 @@ discord-bot = ["agent", "rig-agent/discord-bot"]
 pdf = ["rig-core/pdf"]
 epub = ["rig-core/epub"]
 rayon = ["rig-core/rayon"]
-rmcp = ["agent", "dep:rig-rmcp"]
+rmcp = ["dep:rig-rmcp"]
 socks = ["reqwest", "rig-reqwest/socks"]
 reqwest = ["dep:rig-reqwest"]
 websocket = ["reqwest", "rig-reqwest/websocket"]
