+++ crates/rig-core/src/tool/mod.rs
@@ -4,9 +4,11 @@
 //! without importing a registry, mutable context, lifecycle state, or executor.
 
 pub mod builtin;
+pub mod managed;
 mod output;
 pub mod portable;
 mod result;
+pub use managed::{ManagedToolSink, ManagedToolToken};
 pub use output::{IntoToolOutput, ToolOutput};
 pub use portable::{
     PortableDynamicTool, PortableTool, PortableToolEmbedding, portable_tool_definition,
