+++ crates/rig-rmcp/README.md
@@ -1,8 +1,11 @@
 # rig-rmcp
 
-MCP (Model Context Protocol) tool support for [Rig](https://crates.io/crates/rig), built on the [`rmcp`](https://docs.rs/rmcp) SDK.
+MCP (Model Context Protocol) tool support for [Rig](https://crates.io/crates/rig), built on the [`rmcp`](https://docs.rs/rmcp) SDK. Depends on **rig-core only**.
 
-- With default features (`agent`): `McpTool` implements rig-agent's contextual `ErasedTool` (MCP `_meta` passthrough from the `ToolContext`, raw results preserved on it), `McpClientHandler` keeps a `ToolServer` in sync with `notifications/tools/list_changed`, and `RmcpAgentBuilderExt` / `RmcpToolServerExt` add `rmcp_tool(s)` to the agent and tool-server builders. The `rig` facade re-exports all of it as `rig::tool::rmcp` behind its `rmcp` feature.
-- With `default-features = false`: depends on rig-core only. MCP tools are exposed as `rig_core::tool::PortableDynamicTool`s (no `_meta` passthrough, no preserved raw result — there is no context to put them on), usable from any runtime that consumes portable tools.
+- `McpTool` — one MCP server tool; convert it into a `rig_core::tool::PortableDynamicTool` (`From`) and register it wherever portable tools go (`rig-agent`'s `AgentBuilder::portable_dynamic_tool`, `ToolServer::portable_dynamic_tool`, or any other runtime). The portable tool carries a liveness probe bound to the MCP transport, so registries can retire it on disconnect. `tools_from_server` converts a whole tool list.
+- `McpClientHandler<S: rig_core::tool::ManagedToolSink>` — an rmcp client handler that keeps a tool registry in sync with `notifications/tools/list_changed`. rig-agent's `ToolServerHandle` implements `ManagedToolSink`, so `McpClientHandler::new(client_info, tool_server_handle.clone()).connect(transport)` is the agent usage; other runtimes implement the sink for their own registry.
+- Calls go out without MCP `_meta`; a caller that needs to attach one drives `McpTool::execute_mcp(args, Some(meta))` directly.
+
+The `rig` facade re-exports this crate as `rig::tool::rmcp` behind its `rmcp` feature.
 
 Native-only: rmcp's `ClientHandler` requires `Send + Sync` unconditionally, which rig's wasm tool registry cannot satisfy; building for `wasm32` fails with a single explanatory `compile_error!`.
